// src/models/output.ts
export interface Output {
    date: string
    
    // fajr_time_range: string
    fajr_time: string
    // fajr_angle: number

    sunrise_time: string

    duhur_time: string

    asr_time: string
    asr_hanafi_time: string

    // maghreb_time_range: string
    maghreb_time: string

    // isha_angle: string
    // isha_time_range: string
    isha_time: string
}